src/test/java/com/mrn/didemo/controllers/SetterInjectedControllerTest.java
package com.mrn.didemo.controllers;

import com.mrn.didemo.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class SetterInjectedControllerTest {

    private SetterInjectedController setterInjectedController;

    // new instance of the controller
    // set the controller
    @Before
    public void setUp() {
        this.setterInjectedController = new SetterInjectedController();
        this.setterInjectedController.setGreetingService(new GreetingServiceImpl());
    }

    @Test
    public void testGreeting() {
        assertEquals(GreetingServiceImpl.HIELO_ME, setterInjectedController.sayHello());
    }
}
